Chore/remove unused and update packages (#1027)
* chore: remove unused packages

* chore: upgrade bug-fixes packages

* chore: upgrade [yellow] minor backward-compatible features packages

* chore: upgrade lru-cache in lambdas

* chore: update sinon

* chore: update eslint

* chore: update testcontainers

* chore: update @types/node

* chore: move common packages between content/lambdas to root package.json + sort package.json entries

* chore: update sharp

* chore: update lru-cache types

* chore: update chalk

* chore: update react types

* chore: small fixes/refactors in configs

* chore: update catalyst-node-commons

* chore: restore previous version of chalk because it breaks jest

* chore: use dynamic import for testcontainers

* chore: move static var to instance var

* test: test CI without testcontainers

* test: test if testcontainers is the conflict import

* test: test dynamic import

* chore: re-add nyc as it is needed for reporting api coverage

* test: test if import wait-strategy breaks CI too

* test: test transpile jest setup

* chore: fix testcontainers

* chore: lambdas/content with its dependencies in package.json. added yarn scripts to list-mismatches between lambdas/content dep versions, sort package.json and list unused

* chore: update content-validator and dcl-schemas

* chore: remove @types/node-fetch

* chore: recreate yarn.lock

* chore: add comment for testcontainers dynamically imported

* chore: run list-mismatches on yarn build
