+++ content/jest.config.js
@@ -7,22 +7,22 @@ module.exports = {
   verbose: true,
   projects: [
     {
+      displayName: 'unit',
       globals: {
         'ts-jest': {
-          tsconfig: '<rootDir>/test/tsconfig.json'
+          tsconfig: '<rootDir>/test/tsconfig.json',
         }
       },
-      displayName: 'unit',
       testMatch: ['**/test/unit/**/*.spec.(ts)'],
       preset: 'ts-jest'
     },
     {
+      displayName: 'integration',
       globals: {
         'ts-jest': {
           tsconfig: '<rootDir>/test/tsconfig.json'
         }
       },
-      displayName: 'integration',
       testMatch: ['**/test/integration/**/*.spec.(ts)'],
       globalSetup: './jest.globalSetup.ts',
       globalTeardown: './jest.globalTeardown.ts',
