+++ content/jest.globalTeardown.ts
@@ -2,9 +2,8 @@ import printApiCoverage from './printApiCoverage'
 
 const globalTeardown = async (): Promise<void> => {
   if (process.env.API_COVERAGE === 'true') {
-    await printApiCoverage()
+    printApiCoverage()
   }
-
   await globalThis.__POSTGRES_CONTAINER__?.stop()
 }
