+++ content/package.json
@@ -1,19 +1,19 @@
 {
   "name": "@catalyst/content-server",
-  "description": "Content server",
   "version": "0.1.0",
-  "author": "Decentraland Contributors",
+  "description": "Content server",
   "license": "Apache-2.0",
+  "author": "Decentraland Contributors",
   "scripts": {
-    "cleanup": "shx rm -rf dist node_modules",
     "build": "tsc -b",
-    "start:server": "node ./dist/src/entrypoints/run-server.js",
+    "cleanup": "shx rm -rf dist node_modules",
+    "report-api-coverage": "node ./reportApiCoverage.js",
     "start:db": "node ./dist/src/entrypoints/run-local-database.js",
     "start:maintenance": "node ./dist/src/entrypoints/run-maintenance.js",
+    "start:server": "node ./dist/src/entrypoints/run-server.js",
     "test": "jest --runInBand --forceExit --detectOpenHandles --coverage",
-    "test:unit": "jest --selectProjects unit --runInBand --forceExit --detectOpenHandles --coverage",
     "test:integration": "jest --selectProjects integration --runInBand --forceExit --detectOpenHandles --coverage",
-    "report-api-coverage": "node ./reportApiCoverage.js"
+    "test:unit": "jest --selectProjects unit --runInBand --forceExit --detectOpenHandles --coverage"
   },
   "dependencies": {
     "@dcl/catalyst-api-specs": "1.2.2",
@@ -24,34 +24,31 @@
     "@dcl/schemas": "4.7.0",
     "@dcl/snapshots-fetcher": "3.0.2",
     "@dcl/urn-resolver": "1.3.1",
+    "@well-known-components/http-server": "^1.0.0",
     "@well-known-components/interfaces": "^1.1.1",
     "@well-known-components/logger": "^1.1.0",
-    "aws-sdk": "2.1006.0",
     "bloom-filters": "^3.0.0",
     "compression": "1.7.4",
     "cors": "2.8.5",
     "dcl-catalyst-client": "^11.3.0",
     "dcl-catalyst-commons": "^8.2.0",
     "dcl-crypto": "2.3.0",
-    "eth-crypto": "2.1.0",
-    "express": "4.17.1",
-    "express-openapi-validator": "4.13.1",
-    "form-data": "4.0.0",
+    "eth-crypto": "2.3.0",
+    "express": "4.18.0",
+    "express-openapi-validator": "4.13.7",
     "fp-future": "^1.0.1",
-    "log4js": "6.4.0",
-    "lru-cache": "7.3.1",
+    "log4js": "6.4.6",
+    "lru-cache": "7.8.1",
     "morgan": "1.10.0",
     "ms": "2.1.3",
     "multer": "^1.4.4",
     "node-cache": "^5.1.2",
     "node-fetch": "2.6.7",
-    "node-pg-migrate": "6.0.0",
+    "node-pg-migrate": "6.2.1",
     "p-queue": "6.6.2",
-    "parallel-transform": "1.2.0",
-    "pg": "^8.7.1",
-    "pg-promise": "10.11.0",
+    "pg": "8.7.3",
+    "pg-promise": "10.11.1",
     "pg-query-stream": "^4.2.1",
-    "sharp": "0.29.2",
     "sql-template-strings": "^2.2.2",
     "uuid": "8.3.2"
   },
@@ -61,26 +58,41 @@
     "@types/destroy": "1.0.0",
     "@types/express": "4.17.13",
     "@types/faker": "5.5.8",
-    "@types/jest": "27.4.0",
-    "@types/lru-cache": "5.1.1",
+    "@types/jest": "27.4.1",
+    "@types/lru-cache": "7.6.1",
     "@types/morgan": "^1",
     "@types/multer": "^1",
-    "@types/node": "16.7.10",
-    "@types/node-fetch": "3.0.3",
+    "@types/node": "17.0.29",
     "@types/on-finished": "2.3.1",
-    "@types/parallel-transform": "^1",
     "@types/pg": "^8",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "@types/sinon": "10.0.11",
     "@types/uuid": "^8",
+    "@typescript-eslint/eslint-plugin": "5.21.0",
+    "@typescript-eslint/parser": "5.21.0",
     "chalk": "4.1.2",
     "columnify": "1.6.0",
-    "destroy": "1.0.4",
+    "destroy": "1.2.0",
+    "eslint": "8.14.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-plugin-prettier": "4.0.0",
     "faker": "5.5.3",
     "jest": "27.4.7",
     "jest-extra-utils": "^0.1.0",
     "jest-junit": "^13.0.0",
+    "lint-staged": "12.4.1",
+    "npm-run-all": "4.1.5",
     "nyc": "^15.1.0",
-    "on-finished": "2.3.0",
-    "testcontainers": "2.8.1",
-    "ts-jest": "27.1.2"
+    "on-finished": "2.4.1",
+    "prettier": "2.6.2",
+    "shx": "0.3.4",
+    "sinon": "13.0.2",
+    "testcontainers": "8.9.0",
+    "ts-jest": "27.1.4",
+    "ts-mockito": "2.6.1",
+    "ts-node": "10.7.0",
+    "tsc": "2.0.4",
+    "typescript": "^4.4.3"
   }
 }
