+++ lambdas/package.json
@@ -1,12 +1,12 @@
 {
   "name": "@catalyst/lambdas-server",
-  "description": "Lambdas server",
   "version": "0.1.0",
-  "author": "Decentraland Contributors",
+  "description": "Lambdas server",
   "license": "Apache-2.0",
+  "author": "Decentraland Contributors",
   "scripts": {
-    "cleanup": "shx rm -rf dist node_modules",
     "build": "tsc -b",
+    "cleanup": "shx rm -rf dist node_modules",
     "start": "node ./dist/src/entrypoints/run-server.js",
     "test": "jest --runInBand --forceExit --coverage"
   },
@@ -16,25 +16,49 @@
     "@dcl/catalyst-node-commons": "1.0.3",
     "@dcl/schemas": "4.7.0",
     "@dcl/urn-resolver": "1.3.1",
-    "@types/sharp": "0.29.2",
     "compression": "1.7.4",
     "cors": "2.8.5",
-    "dcl-catalyst-client": "^11.3.0",
     "dcl-catalyst-commons": "^8.2.0",
     "dcl-crypto": "2.3.0",
-    "express": "4.17.1",
-    "express-openapi-validator": "4.13.1",
-    "fp-future": "1.0.1",
-    "log4js": "6.4.0",
-    "lru-cache": "6.0.0",
+    "express": "4.18.0",
+    "express-openapi-validator": "4.13.7",
+    "fp-future": "^1.0.1",
+    "log4js": "6.4.6",
+    "lru-cache": "7.8.1",
     "morgan": "1.10.0",
     "ms": "2.1.3",
     "node-fetch": "2.6.7",
-    "sharp": "0.29.1"
+    "sharp": "0.30.4"
   },
   "devDependencies": {
-    "@types/jest": "27.4.0",
+    "@types/compression": "^1",
+    "@types/cors": "^2",
+    "@types/express": "4.17.13",
+    "@types/jest": "27.4.1",
+    "@types/lru-cache": "7.6.1",
+    "@types/morgan": "^1",
+    "@types/node": "17.0.29",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "@types/sharp": "0.30.2",
+    "@types/sinon": "10.0.11",
+    "@types/uuid": "^8",
+    "@typescript-eslint/eslint-plugin": "5.21.0",
+    "@typescript-eslint/parser": "5.21.0",
+    "eslint": "8.14.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-plugin-prettier": "4.0.0",
     "jest": "27.4.7",
-    "ts-jest": "27.1.2"
+    "jest-junit": "^13.0.0",
+    "lint-staged": "12.4.1",
+    "npm-run-all": "4.1.5",
+    "prettier": "2.6.2",
+    "shx": "0.3.4",
+    "sinon": "13.0.2",
+    "ts-jest": "27.1.4",
+    "ts-mockito": "2.6.1",
+    "ts-node": "10.7.0",
+    "tsc": "2.0.4",
+    "typescript": "^4.4.3"
   }
 }
