+++ package.json
@@ -1,27 +1,34 @@
 {
   "name": "catalyst",
-  "private": true,
   "version": "0.1.0",
-  "author": "Decentraland Contributors",
+  "private": true,
   "license": "Apache-2.0",
+  "author": "Decentraland Contributors",
   "workspaces": [
     "content",
     "lambdas"
   ],
   "scripts": {
+    "build": "yarn list-mismatches && yarn workspaces run build",
+    "build:content": "yarn list-mismatches && yarn workspace @catalyst/content-server build",
+    "build:lambdas": "yarn list-mismatches && yarn workspace @catalyst/lambdas-server build",
     "cleanup": "yarn workspaces run cleanup && shx rm -rf node_modules",
-    "build": "yarn workspaces run build",
-    "build:lambdas": "yarn workspace @catalyst/lambdas-server build",
-    "build:content": "yarn workspace @catalyst/content-server build",
-    "start": "yarn workspaces run start",
-    "start:lambdas": "yarn workspace @catalyst/lambdas-server start",
-    "start:content": "yarn workspace @catalyst/content-server start:db && yarn workspace @catalyst/content-server start:server",
-    "test": "run-s 'test:*'",
-    "test:lambdas": "yarn workspace @catalyst/lambdas-server test",
-    "test:content": "yarn workspace @catalyst/content-server test",
-    "lint:fix": "eslint '**/*.{js,ts}' --fix",
     "lint:check": "eslint '**/*.{js,ts}'",
-    "migrate": "node-pg-migrate -m content/src/migrations/scripts -j ts"
+    "lint:fix": "eslint '**/*.{js,ts}' --fix",
+    "list-mismatches": "npx syncpack list-mismatches --source 'package.json' --source 'lambdas/package.json' --source 'content/package.json'",
+    "migrate": "node-pg-migrate -m content/src/migrations/scripts -j ts",
+    "sort-pkg": "npx sort-package-json; yarn sort-pkg:content; yarn sort-pkg:lambdas",
+    "sort-pkg:content": "npx sort-package-json content/package.json",
+    "sort-pkg:lambdas": "npx sort-package-json lambdas/package.json",
+    "start": "yarn workspaces run start",
+    "start:content": "yarn workspace @catalyst/content-server start:db && yarn workspace @catalyst/content-server start:server",
+    "start:lambdas": "yarn workspace @catalyst/lambdas-server start",
+    "test": "run-s 'test:*'",
+    "test:content": "yarn workspace @catalyst/content-server test",
+    "test:lambdas": "yarn workspace @catalyst/lambdas-server test",
+    "unused": "npx depcheck; yarn unused:content; yarn unused:lambdas",
+    "unused:content": "npx depcheck content",
+    "unused:lambdas": "npx depcheck lambdas"
   },
   "lint-staged": {
     "**/*.{js,ts,tsx}": [
@@ -35,32 +42,12 @@
     "trailingComma": "none"
   },
   "devDependencies": {
-    "@types/cors": "^2",
-    "@types/express": "4.17.13",
-    "@types/morgan": "^1",
-    "@types/node": "16.7.10",
-    "@types/react": "^17",
-    "@types/react-dom": "^17",
-    "@types/sinon": "10.0.4",
-    "@types/uuid": "^8",
-    "@typescript-eslint/eslint-plugin": "4.33.0",
-    "@typescript-eslint/parser": "4.33.0",
-    "cpy-cli": "^4.1.0",
-    "eslint": "7.32.0",
-    "eslint-config-prettier": "8.3.0",
+    "@typescript-eslint/eslint-plugin": "5.21.0",
+    "@typescript-eslint/parser": "5.21.0",
+    "eslint": "8.14.0",
     "eslint-plugin-prettier": "4.0.0",
-    "lint-staged": "12.1.4",
-    "mock-socket": "8.0.5",
-    "node-pg-migrate": "6.0.0",
-    "npm-run-all": "4.1.5",
-    "pg": "8.7.1",
-    "prettier": "2.4.1",
-    "shx": "0.3.3",
-    "sinon": "11.1.2",
-    "ts-mockito": "2.6.1",
-    "ts-node": "10.2.1",
-    "ts-path-mapping": "0.1.4",
-    "tsc": "2.0.3",
+    "lint-staged": "12.4.1",
+    "prettier": "2.6.2",
     "typescript": "^4.4.3"
   },
   "jest-junit": {
