+++ tsconfig.json
@@ -17,7 +17,6 @@
     "noUnusedLocals": true,
     "jsx": "react",
     "skipLibCheck": true,
-    "outDir": "./dist",
     "lib": [
       "es2020",
       "dom",
